Add possibility to add mutiple project_type conf

+++ config/app/bin/script-symfony-flex.sh
@@ -11,7 +11,10 @@ chmod 777 \
     ${APP_ROOT}/${APP_LOGS_DIR} \
     ${APP_ROOT}/${APP_SESSIONS_DIR}
 
-rm -f /usr/local/etc/php/conf.custom.d/*-a_p_t-*.ini
+if [[ -z ${APP_ROOT_MULTIPLE}  ]]
+then
+    rm -f /usr/local/etc/php/conf.custom.d/*-a_p_t-*.ini
+fi
 cp -f /usr/local/etc/php/conf.project_type.d/*${APP_PROJECT_TYPE}.ini /usr/local/etc/php/conf.custom.d/
 
 exec "$@"
